// src/db/database.types.ts
export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          extensions?: Json;
          operationName?: string;
          query?: string;
          variables?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          created_at: string;
          deleted_at: string | null;
          id: string;
          username: string;
        };
        Insert: {
          created_at?: string;
          deleted_at?: string | null;
          id: string;
          username: string;
        };
        Update: {
          created_at?: string;
          deleted_at?: string | null;
          id?: string;
          username?: string;
        };
        Relationships: [];
      };
      prompt_tags: {
        Row: {
          prompt_id: string;
          tag_id: string;
        };
        Insert: {
          prompt_id: string;
          tag_id: string;
        };
        Update: {
          prompt_id?: string;
          tag_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "prompt_tags_prompt_id_fkey";
            columns: ["prompt_id"];
            isOneToOne: false;
            referencedRelation: "prompts";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "prompt_tags_tag_id_fkey";
            columns: ["tag_id"];
            isOneToOne: false;
            referencedRelation: "tags";
            referencedColumns: ["id"];
          },
        ];
      };
      prompts: {
        Row: {
          author_id: string;
          content: string;
          created_at: string;
          deleted_at: string | null;
          description: string | null;
          id: string;
          title: string;
          updated_at: string | null;
        };
        Insert: {
          author_id: string;
          content: string;
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          title: string;
          updated_at?: string | null;
        };
        Update: {
          author_id?: string;
          content?: string;
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          title?: string;
          updated_at?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "prompts_author_id_fkey";
            columns: ["author_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
        ];
      };
      tags: {
        Row: {
          created_at: string;
          id: string;
          name: string;
        };
        Insert: {
          created_at?: string;
          id?: string;
          name: string;
        };
        Update: {
          created_at?: string;
          id?: string;
          name?: string;
        };
        Relationships: [];
      };
      votes: {
        Row: {
          created_at: string;
          id: string;
          prompt_id: string;
          user_id: string;
          vote_value: number;
        };
        Insert: {
          created_at?: string;
          id?: string;
          prompt_id: string;
          user_id: string;
          vote_value: number;
        };
        Update: {
          created_at?: string;
          id?: string;
          prompt_id?: string;
          user_id?: string;
          vote_value?: number;
        };
        Relationships: [
          {
            foreignKeyName: "votes_prompt_id_fkey";
            columns: ["prompt_id"];
            isOneToOne: false;
            referencedRelation: "prompts";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: Record<never, never>;
    Functions: Record<never, never>;
    Enums: {
      event_type: "'prompt_view'" | "'prompt_copy'";
      flag_reason: "'inaccurate'" | "'outdated'" | "'unclear'";
    };
    CompositeTypes: Record<never, never>;
  };
}